requirements.txt
aiogram==2.19
aiogram-calendar==0.1.1
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
BigQuery-Python==1.15.0
cachetools
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
cycler==0.11.0
db-dtypes==1.0.0
Deprecated==1.2.13
distlib==0.3.4
filelock==3.6.0
fonttools==4.33.3
frozenlist==1.3.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
identify==2.5.0
idna==3.3
iniconfig==1.1.1
kiwisolver==1.4.2
matplotlib==3.5.1
multidict==6.0.2
nodeenv==1.6.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pluggy==1.0.0
pre-commit==2.18.1
proto-plus==1.20.3
protobuf==3.20.1
py==1.11.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
wrapt==1.14.0
yarl==1.7.2
https://github.com/Birdi7/inline-timepicker/archive/master.zip
Jinja2==3.1.2